Fixed bug #13871: Field data_page_type in table index_phash is too small.
git-svn-id: https://svn.typo3.org/TYPO3v4/Core/trunk@7137 709f56b5-9817-0410-a4d7-c38de5d9e867

+++ typo3/sysext/indexed_search/ext_tables.sql
@@ -9,7 +9,7 @@ CREATE TABLE index_phash (
   data_filename varchar(255) DEFAULT '' NOT NULL,
   data_page_id int(11) unsigned DEFAULT '0' NOT NULL,
   data_page_reg1 int(11) unsigned DEFAULT '0' NOT NULL,
-  data_page_type tinyint(3) unsigned DEFAULT '0' NOT NULL,
+  data_page_type int(11) unsigned DEFAULT '0' NOT NULL,
   data_page_mp varchar(255) DEFAULT '' NOT NULL,
   gr_list varchar(255) DEFAULT '' NOT NULL,
   item_type varchar(5) DEFAULT '' NOT NULL,
